Allow unzipped bundled plugins for JIRA
The code here parallels the confluence implementation of the same feature. It
is driven by the presence of the expanded directory rather than a specific version.
Also includes some sharing of common strings with tests, plus tests for the new
functionality, more generic cleanup in the tests to cope with new test directories, and
a detabbing fix.

For AMPS-1086.

Conflicts:
	maven-amps-plugin/src/main/java/com/atlassian/maven/plugins/amps/product/JiraProductHandler.java
	maven-amps-plugin/src/test/java/com/atlassian/maven/plugins/amps/product/TestJiraProductHandler.java