Mark jira4 and jira5 wrapper scripts deprecated
Jira 4 has reached end of support and so atlas-create-jira4-plugin has
been deprecated. The atlassian-create-jira5-plugin script actually sets
the jira.version to 6.3 anyway, so let's deprecate it and tell people to
edit the version directory, which is how the other products expect you
to work.

I added support for declaring deprecation to the velocity macros which
generate these scripts. The .bat versions were tested with wine under
OSX, the sh versions using dash on OSX.

For AMPS-1156