AMPSDEV-183 Remove MaxPermSize from JvmArgsFix
The permgen was removed in Java 8 (see JEP 122) and since AMPS 6.1
onwards requires Java 8, we can and should remove tuning of perm gen.
The in Java 8 the JVM command line argument -XX:MaxPermSize is ignored,
and results in a warning being logged. In Java 9, at least for the
current EA builds, the JVM will not start when this argument is passed.