AMPS-1439 Fix configuration for executed mojos.
org.twdata.maven:mojo-executor:1.5.2, used by AMPS 6.3, automatically
merged project-level configuration and AMPS-provided configuration as
part of executing mojos. 2.0.0 and newer (including 2.3.0, currently
in use) removed that merging, and now use the provided configuration
exclusively.

This effectively broke amps-dispatcher-maven-plugin, as it meant the
product version (and other product plugin configuration) from the POM
was ignored and the "latest" version of the product was always used.

It also had more subtle knock-on effects, like breaking the ability to
apply configuration to the maven-surefire-plugin and have that merged
with (and potentially override) the AMPS-provided default configuration.

- Added MojoUtils, which wraps MojoExecutor and offers two methods for
  invoking a mojo:
  - execute: New-style MojoExecutor.executeMojo behavior, where only the
    provided configuration is used
  - executeWithMergedConfig: Old-style MojoExecutor.executeMojo behavior
    where the provided configuration serves as _defaults_ and can be
    overridden by project-level configuration in the POM(s)
- Updated all MojoExecutor.executeMojo calls to use MojoUtils instead
  - execute is used for running Cargo, to ensure any project-level Cargo
    configuration is ignored
    - This obsoletes MavenGoals.executeMojoExcludeProductCargoConfig
  - executeWithMergedConfig is used everywhere else