AMPS-1468 - Allow developers to set additional Closure Compiler options independent of AMPS being updated.
Developers can now set the output language level, as well as whether strict mode is generated or not. These options are the more important ones for people to affect.

Let's say that it's 3 years from now and the minimum language level for JS in our supported browser matrix is ECMASCRIPT_2018. We might not have updated AMPS by the time that happens. Rather than forcing them to abandon AMPS, I'd rather they have the option to change the output language level.

We will still need to update Google Closure every year or so, though, otherwise we'll not be able to support ECMASCRIPT_2021.

There may be a better way to provide a hook to a minifier without AMPS having to know about it... but it's beyond the scope of this refactor and this story.